vissE: Visualising Set Enrichment Analysis Results

This package enables the interpretation and analysis of results from a gene set enrichment analysis using network-based and text-mining approaches. Most enrichment analyses result in large lists of significant gene sets that are difficult to interpret. Tools in this package help build a similarity-based network of significant gene sets from a gene set enrichment analysis that can then be investigated for their biological function using text-mining approaches.

Details

This package supports four workflows to enhance gene set enrichment analysis:

Clustering results from a gene set enrichment analysis (e.g. using limma::fry, singscore or GSEA). The functions required for this analysis are computeMsigOverlap, computeMsigNetwork and plotMsigNetwork.

Interpreting gene set clusters (identified in the first analysis) by performing text-mining of gene set names and descriptions. The main function required to perform text-mining of gene sets is plotMsigWordcloud. Other functions can be used to access intermmediate results.

Visualise gene-level statistics for gene set clusters identified in the first analysis to link back gene set clusters to the genes of interest. This can be done using the plotGeneStats function.

Identifying gene sets similar to a list of genes identified from a DE analysis using set overlap measures. This can be done using the characteriseGeneset function.

See also

Author

Maintainer: Dharmesh D. Bhuva bhuva.d@wehi.edu.au (ORCID)